Separate cross-browser visdiffs (#256)
* Revamped visdiff tests [skip ci]

* Told CircleCI to ignore *visdiff* branches

* Added browserName to top level of spec output, visdiff trigger script

* Renamed multi-browser option in run.sh from -d to -m

* Set grunt path in run.sh

* Removed browserName from visdiff spec output due to timing of when it's generated

* Updated devdocs element link selector for wp-calypso/7866 [skip ci]

* Moved setting batchName into main .before() call to have browserName available

* Updated batch/test name with 'visdiff'
