+++ circle.yml
@@ -9,6 +9,7 @@ general:
   branches:
     ignore:
       - /.*[j,J]etpack.*/ # ignore branches with jetpack in their name for Circle - these are run in a separate build
+      - /.*[v,V]isdiff.*/ # ignore branches with visdiff in their name for Circle - these are run in a separate build
 
 dependencies:
   pre:
