+++ lib/pages/devdocs-page.js
@@ -1,50 +1,21 @@
 import webdriver from 'selenium-webdriver';
-
+import * as driverHelper from '../driver-helper.js'
 import BaseContainer from '../base-container.js';
 
 const by = webdriver.By;
 
-export default class DevdocsDesignPage extends BaseContainer {
+export default class DevdocsPage extends BaseContainer {
 	constructor( driver, visit ) {
 		const baseURL = 'https://wpcalypso.wordpress.com/devdocs';
 		//const baseURL = 'http://calypso.localhost:3000/devdocs';
 
 		super( driver, by.css( '.devdocs.main' ), visit, baseURL );
 
-		// The devdocs page is transitioning to a new standard classname
-		let oldClassName = '.design-assets__group';
-		let newClassName = '.docs__design-assets-group';
-
 		this.baseURL = baseURL;
-		this.designElementSelector = by.css( `${oldClassName}, ${newClassName}` );
-		this.designElementLinkSelector = by.css( `${oldClassName} h2 a:not(.button), ${newClassName} h2 a:not(.button)` );
+		this.designElementLinkSelector = by.css( '.docs-example__wrapper-header-title:not(.button)' );
 		this.elementTitleSelector = by.css( '.header-cake__title' );
-		this.elementButtonSelector = by.css( `div:not([style*="display:none"]):not([style*="display: none"]) > ${oldClassName} h2 a.button, div:not([style*="display:none"]):not([style*="display: none"]) > ${newClassName} h2 a.button` );
+		this.elementButtonSelector = by.css( '.docs__design-toggle.button' );
 		this.allComponentsSelector = by.css( 'button.header-cake__back' );
-	}
-
-	compactAllElements() {
-		const selector = '.design-assets__toggle.button';
-		var driver = this.driver;
-
-		var d = webdriver.promise.defer();
-
-		driver.findElements( by.css( selector ) ).then( function( buttons ) {
-			let promiseArray = [];
-
-			for ( let i = 0; i < buttons.length; i++ ) {
-				promiseArray.push( buttons[i].click() );
-			}
-
-			webdriver.promise.all( promiseArray ).then( function() {
-				// Scroll back to the top of the page
-				driver.executeScript( 'window.scrollTo( 0, 0 )' ).then( function() {
-					d.fulfill( true );
-				} );
-			} );
-		} );
-
-		return d.promise;
 	}
 
 	openUIComponents() {
@@ -68,12 +39,7 @@ export default class DevdocsDesignPage extends BaseContainer {
 		} );
 	}
 
-	getAllDesignElements() {
-		return this.driver.findElements( this.designElementSelector );
-	}
-
 	getAllDesignElementLinks() {
-//		return this.driver.findElements( this.designElementLinkSelector );
 		var hrefs = [];
 		var d = webdriver.promise.defer();
 
@@ -122,5 +88,56 @@ export default class DevdocsDesignPage extends BaseContainer {
 
 	hideEnvironmentBadge() {
 		return this.driver.executeScript( 'document.querySelector( ".environment-badge" ).style.visibility = "hidden";' );
+	}
+
+	screenshotAllElements( eyes ) {
+		const self = this;
+		const driver = this.driver;
+
+		return this.getAllDesignElementLinks().then( function( hrefs ) {
+			let flow = driver.controlFlow();
+
+			for ( const href of hrefs ) {
+				let title;
+
+				// Open the design element
+				flow.execute( function() {
+					let titleSplit = href.split( '/' );
+					title = titleSplit[ titleSplit.length - 1 ];
+					return driver.get( href );
+				} );
+				// Hide the masterbar for clean CSS stitching
+				flow.execute( function() {
+					return self.hideMasterbar();
+				} );
+				// Hide the environment badge for simplicity
+				flow.execute( function() {
+					return self.hideEnvironmentBadge();
+				} );
+				// Take the screenshot
+				flow.execute( function() {
+					return driverHelper.eyesScreenshot( driver, eyes, title, by.id( 'primary' ) );
+				} );
+				// Check for Compact button
+				flow.execute( function() {
+					return self.isCurrentElementCompactable().then( function( compactable ) {
+						if ( compactable ) {
+							return self.getCurrentElementCompactButton().then( function( button ) {
+								// Some browsers need a more precise click on the mobile width to avoid overlapping elements
+								if ( global.browserName.toLowerCase() === 'chrome' || global.browserName.toLowerCase() === 'firefox' ) {
+									return driver.actions().mouseMove( button, {x: 3, y: 3} ).click().perform().then( function() {
+										return driverHelper.eyesScreenshot( driver, eyes, title + ' (Compact)', by.id( 'primary' ) );
+									} );
+								}
+
+								return button.click().then( function() {
+									return driverHelper.eyesScreenshot( driver, eyes, title + ' (Compact)', by.id( 'primary' ) );
+								} );
+							} );
+						}
+					} );
+				} );
+			}
+		} );
 	}
 }
